samsung_auto_trader/requirements.txt
requests
PyYAML
pandas
pycryptodome
websockets
